refactor: simplify run scripts and create exchange if it not exists
